requirements.txt
newspaper3k==0.2.8
notebook==6.1.4
spacy==2.3.2
nltk==3.4.4
numpy==1.19.2
pandas==1.1.4
es_core_news_md==2.3.1
beautifulsoup4==4.9.3
